+++ parser_library/src/context/id_storage.cpp
@@ -69,10 +69,10 @@ id_storage::well_known_strings::well_known_strings(std::unordered_set<std::strin
     , MACRO(&*ptr.emplace("MACRO").first)
     , MEND(&*ptr.emplace("MEND").first)
     , MEXIT(&*ptr.emplace("MEXIT").first)
+    , MHELP(&*ptr.emplace("MHELP").first)
     , ASPACE(&*ptr.emplace("ASPACE").first)
     , AIF(&*ptr.emplace("AIF").first)
     , AGO(&*ptr.emplace("AGO").first)
     , ACTR(&*ptr.emplace("ACTR").first)
     , AREAD(&*ptr.emplace("AREAD").first)
-    , empty(&*ptr.emplace("").first)
 {}