+++ parser_library/src/diagnostic.cpp
@@ -1142,6 +1142,14 @@ diagnostic_op diagnostic_op::warning_A249_sequence_symbol_expected(const range& 
     return diagnostic_op(diagnostic_severity::warning, "A249", "Sequence symbol expected", range);
 }
 
+diagnostic_op diagnostic_op::error_A250_absolute_with_known_symbols(const range& range)
+{
+    return diagnostic_op(diagnostic_severity::error,
+        "A250",
+        "Operand must be an absolute expression of previously defined symbols",
+        range);
+}
+
 diagnostic_op diagnostic_op::warning_A300_op_apostrophes_missing(const std::string& instr_name, const range& range)
 {
     return diagnostic_op(diagnostic_severity::warning,
@@ -1826,6 +1834,17 @@ diagnostic_op diagnostic_op::error_E070(const range& range)
 diagnostic_op diagnostic_op::error_E071(const range& range)
 {
     return diagnostic_op(diagnostic_severity::error, "E071", "Macro prototype expected.", range);
+}
+
+diagnostic_op diagnostic_op::error_E072(const range& range)
+{
+    return diagnostic_op(diagnostic_severity::error, "E072", "SYSNDX limit reached, macro call supressed.", range);
+}
+
+diagnostic_op diagnostic_op::error_E073(const range& range)
+{
+    return diagnostic_op(
+        diagnostic_severity::error, "E073", "Illegal START instruction - CSECT already exists.", range);
 }
 
 diagnostic_op diagnostic_op::warning_W010(const std::string& message, const range& range)